render: Change the Render trait up.
This removes the Render implementation for the member event but adds it
to a bunch of contents.

Turns out we need to render the contents without a prefix for local
echo, so we need to change the trait up a bit.

At the same time we'll never need to render member events without a
prefix, so it makes sense to special case them.