room: Smarter handling for prev_batch
Now that we restore our rooms again we need to be a bit smarter about
our prev_batches. If our buffer is empty we'll need to fetch events in
the froward direction the first time, on the other hand if we receive
a sync response and our buffer is empty we need to set it to the new
prev_batch token.