fix: remove #[used] from non-static fields
Rust 1.51 has a regression restricting the `#[used]` attribute[0] to
static fields. As such, it's not possible to build `weechat-matrix-rs`
on stable or nightly Rust. According to the `#[used]` RFC[1], the
attribute is for static variables so it's unclear if the new behavior is
the correct one.

However, to prevent clippy from complaining during CI, replace with
`#[allow(dead_code)]`

[0] https://github.com/rust-lang/rust/issues/82668
[1] https://rust-lang.github.io/rfcs/2386-used.html